Updated the tables to be consistent with the new mindmap
Added the new leafs that are on the new mindmap to the tables. Did not remove any entries from the tables (even the ones that are not on the mindmap).  Some additional notes below.

The new mindmap does not have “3D model creation”. Instead it has “3D Models from Point Clouds”. So I changed the name of the leaf “3D model creation” to “3D Models from Point Clouds”

The new mindmap does not have  “Indoor positioning”. Instead it has “Indoor: Position, Models and Navigation”. So I changed the name of the leaf “Indoor positioning” to “Indoor: Position, Models and Navigation”

Made an observation that “Indoor: Position, Models and Navigation” overlaps with “Indoor venue maps”. Since the new mindmap does not have “Indoor venue maps”, I suggest the description of “Indoor: Position, Models and Navigation” be extended to include that of “Indoor venue maps”.